% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_sort.R
\name{bubble_sort_rust}
\alias{bubble_sort_rust}
\title{Bubble Sorting - Rust Implementation}
\usage{
bubble_sort_rust(n)
}
\arguments{
\item{n}{numeric vector to sort}
}
\value{
a sorted integer vector
}
\description{
Bubble Sorting - Rust Implementation
}
